refactor: make state_store a root module. Split StateStore trait into 2 traits: LoadState and PeristDecision. Hide internal modules and expose public types.
