+++ disintegrate-postgres/src/lib.rs
@@ -9,9 +9,7 @@ pub use crate::event_store::PgEventStore;
 #[cfg(feature = "listener")]
 pub use crate::listener::{PgEventListener, PgEventListenerConfig};
 pub use crate::snapshotter::PgSnapshotter;
-use disintegrate::{
-    DecisionMaker, Event, EventSourcedDecisionStateStore, NoSnapshot, WithSnapshot,
-};
+use disintegrate::{DecisionMaker, Event, EventSourcedStateStore, NoSnapshot, WithSnapshot};
 use disintegrate_serde::Serde;
 pub use error::Error;
 
@@ -19,7 +17,7 @@ pub type PgEventId = i64;
 
 /// An alias for [`DecisionMaker`], specialized for Postgres.
 pub type PgDecisionMaker<E, S, SN> =
-    DecisionMaker<EventSourcedDecisionStateStore<PgEventId, E, PgEventStore<E, S>, SN>>;
+    DecisionMaker<EventSourcedStateStore<PgEventId, E, PgEventStore<E, S>, SN>>;
 
 /// An alias for [`WithSnapshot`], specialized for Postgres.
 pub type WithPgSnapshot = WithSnapshot<PgEventId, PgSnapshotter>;
@@ -45,7 +43,7 @@ pub async fn decision_maker_with_snapshot<
 ) -> Result<PgDecisionMaker<E, S, WithPgSnapshot>, Error> {
     let pool = event_store.pool.clone();
     let snapshot = WithSnapshot::new(PgSnapshotter::new(pool, every).await?);
-    Ok(DecisionMaker::new(EventSourcedDecisionStateStore::new(
+    Ok(DecisionMaker::new(EventSourcedStateStore::new(
         event_store,
         snapshot,
     )))
@@ -63,5 +61,5 @@ pub async fn decision_maker_with_snapshot<
 pub fn decision_maker<E: Event + Send + Sync + Clone, S: Serde<E> + Clone + Sync + Send>(
     event_store: PgEventStore<E, S>,
 ) -> PgDecisionMaker<E, S, NoSnapshot> {
-    DecisionMaker::new(EventSourcedDecisionStateStore::new(event_store, NoSnapshot))
+    DecisionMaker::new(EventSourcedStateStore::new(event_store, NoSnapshot))
 }
